+++ src/combinator/provided.rs
@@ -1,9 +1,11 @@
 mod eat;
 mod next;
+mod take;
 mod till;
 mod wrap;
 
 pub use eat::*;
 pub use next::*;
+pub use take::*;
 pub use till::*;
 pub use wrap::*;