+++ src/combinator/provided/eat.rs
@@ -44,36 +44,48 @@ unsafe impl<State, Heap> Action<str, State, Heap> for Eat<&str> {
   }
 }
 
-unsafe impl<State, Heap> Action<str, State, Heap> for Eat<usize> {
+unsafe impl<State, Heap> Action<[u8], State, Heap> for Eat<u8> {
   type Value = ();
 
   #[inline]
-  fn exec(&self, input: Input<&str, &mut State, &mut Heap>) -> Option<Output<()>> {
-    let mut digested: usize = 0;
-    let mut count: usize = 0;
-    let mut chars = input.instant().rest().chars();
-    while count < self.inner {
-      // no enough chars, try to digest more
-      if let Some(c) = chars.next() {
-        digested = unsafe { digested.unchecked_add(c.len_utf8()) };
-        // SAFETY: count is always smaller than self which is a usize
-        count = unsafe { count.unchecked_add(1) };
-      } else {
-        // no enough chars, reject
-        return None;
-      }
-    }
-    // enough chars
-    unsafe { input.digest_unchecked(digested) }.into()
+  fn exec(&self, input: Input<&[u8], &mut State, &mut Heap>) -> Option<Output<()>> {
+    input
+      .instant()
+      .rest()
+      .get(0)
+      .map_or(false, |&c| c == self.inner)
+      .then(|| unsafe { input.digest_unchecked(1) })
   }
 }
+
+unsafe impl<State, Heap> Action<[u8], State, Heap> for Eat<&[u8]> {
+  type Value = ();
+
+  #[inline]
+  fn exec(&self, input: Input<&[u8], &mut State, &mut Heap>) -> Option<Output<()>> {
+    input
+      .instant()
+      .rest()
+      .starts_with(self.inner)
+      .then(|| unsafe { input.digest_unchecked(self.inner.len()) })
+  }
+}
+
+/// See [`eat`]. This trait is only used for early type checking.
+pub trait EatPattern {}
+impl EatPattern for char {}
+impl EatPattern for String {}
+impl EatPattern for &str {}
+impl EatPattern for u8 {}
+impl EatPattern for &[u8] {}
 
 /// Returns a combinator to eat from the head of [`Instant::rest`](crate::instant::Instant::rest) by the provided pattern.
 /// The combinator will reject if the pattern is not found.
 ///
-/// `0` and `""` (empty string) are allowed but be careful with infinite loops.
+/// `""` (empty string) is allowed but be careful with infinite loops.
 ///
 /// # Examples
+/// For string (`&str`)
 /// ```
 /// # use whitehole::{combinator::{eat, Combinator}, action::Action};
 /// # fn t(_: Combinator<impl Action>) {}
@@ -86,12 +98,20 @@ unsafe impl<State, Heap> Action<str, State, Heap> for Eat<usize> {
 /// # t(
 /// eat("true".to_string()) // eat by String
 /// # );
+/// ```
+/// For bytes (`&[u8]`)
+/// ```
+/// # use whitehole::{combinator::{eat, Combinator}, action::Action};
+/// # fn t(_: Combinator<impl Action<[u8]>>) {}
 /// # t(
-/// eat(10) // eat by char count (not byte length)
+/// eat(b'a') // eat by a byte (u8)
+/// # );
+/// # t(
+/// eat(b"true") // eat by &[u8]
 /// # );
 /// ```
 #[inline]
-pub const fn eat<T>(pattern: T) -> Combinator<Eat<T>> {
+pub const fn eat<T: EatPattern>(pattern: T) -> Combinator<Eat<T>> {
   Combinator::new(Eat::new(pattern))
 }
 
@@ -105,14 +125,18 @@ macro_rules! impl_into_eat_combinator {
     }
   };
 }
-
 impl_into_eat_combinator!(char);
 impl_into_eat_combinator!(String);
-impl_into_eat_combinator!(usize);
-
+impl_into_eat_combinator!(u8);
 impl<'a> From<&'a str> for Combinator<Eat<&'a str>> {
   #[inline]
   fn from(v: &str) -> Combinator<Eat<&str>> {
+    eat(v)
+  }
+}
+impl<'a> From<&'a [u8]> for Combinator<Eat<&'a [u8]>> {
+  #[inline]
+  fn from(v: &[u8]) -> Combinator<Eat<&[u8]>> {
     eat(v)
   }
 }
@@ -124,12 +148,12 @@ mod tests {
 
   #[test]
   fn combinator_eat() {
-    // normal usize
+    // normal char
     assert_eq!(
-      eat(3)
-        .exec(Input::new(Instant::new("123"), &mut (), &mut ()))
+      eat(';')
+        .exec(Input::new(Instant::new(";"), &mut (), &mut ()))
         .map(|output| output.digested),
-      Some(3)
+      Some(1)
     );
     // normal &str
     assert_eq!(
@@ -145,19 +169,19 @@ mod tests {
         .map(|output| output.digested),
       Some(3)
     );
-    // normal char
+    // normal u8
     assert_eq!(
-      eat(';')
-        .exec(Input::new(Instant::new(";"), &mut (), &mut ()))
+      eat(b';')
+        .exec(Input::new(Instant::new(b";"), &mut (), &mut ()))
         .map(|output| output.digested),
       Some(1)
     );
-    // overflow
+    // normal &[u8]
     assert_eq!(
-      eat(3)
-        .exec(Input::new(Instant::new("12"), &mut (), &mut ()))
+      eat(b";" as &[u8])
+        .exec(Input::new(Instant::new(b";"), &mut (), &mut ()))
         .map(|output| output.digested),
-      None
+      Some(1)
     );
     // reject
     assert!(eat("123")
@@ -166,13 +190,6 @@ mod tests {
     assert!(eat('1')
       .exec(Input::new(Instant::new("abc"), &mut (), &mut ()))
       .is_none());
-    // 0 is allowed and always accept
-    assert_eq!(
-      eat(0)
-        .exec(Input::new(Instant::new("123"), &mut (), &mut ()))
-        .map(|output| output.digested),
-      Some(0)
-    );
     // empty string is allowed and always accept
     assert_eq!(
       eat("")
@@ -180,18 +197,11 @@ mod tests {
         .map(|output| output.digested),
       Some(0)
     );
-    // eat by chars not bytes
     assert_eq!(
-      eat(1)
-        .exec(Input::new(Instant::new("好"), &mut (), &mut ()))
+      eat("")
+        .exec(Input::new(Instant::new(""), &mut (), &mut ()))
         .map(|output| output.digested),
-      Some(3)
-    );
-    assert_eq!(
-      eat(2)
-        .exec(Input::new(Instant::new("好好"), &mut (), &mut ()))
-        .map(|output| output.digested),
-      Some(6)
+      Some(0)
     );
   }
 
@@ -200,9 +210,13 @@ mod tests {
     fn test(c: Combinator<impl Action>) {
       c.exec(Input::new(Instant::new("a"), &mut (), &mut ()));
     }
-    test(1.into());
+    fn test_bytes(c: Combinator<impl Action<[u8]>>) {
+      c.exec(Input::new(Instant::new(b"a"), &mut (), &mut ()));
+    }
     test('a'.into());
     test("a".into());
     test("a".to_string().into());
+    test_bytes(b'a'.into());
+    test_bytes((b"a" as &[u8]).into());
   }
 }